requirements.txt
Django==2.2.4
PyYAML==5.1.2
pytz==2019.2
requests-oauthlib==1.2.0